requirements.txt
aiogram
python-vlc
python-dotenv
pymongo
motor
pathvalidate
rich
